Added Pyramids
Pyramids added and tested in isolation.
There is a bug currently where a scene of various primitives seems to
cause unintentional edits to vertices of previous meshes. This is most
likely an indexing error that wasn't caught previously due to the
complexity of the shark model. I'm going to hunt it down today.